Addition of systemctl stop and simplified code
