+++ ansible/full.yml
@@ -19,3 +19,4 @@
   when: tpcc is not defined
 - import_playbook: tpcc.yml
   when: tpcc is defined
+- import_playbook: cleanup.yml
