+++ ansible/inventory.yml
@@ -1,13 +1,3 @@
-# Copyright 2021 The Vitess Authors.
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#    http://www.apache.org/licenses/LICENSE-2.0
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
 
 ---
 all:
@@ -64,6 +54,22 @@ all:
               port: 15002
               mysql_port: 13307
               grpc_port: 15307
+            - id: 3
+              port: 15003
+              mysql_port: 13308
+              grpc_port: 15308
+            - id: 4
+              port: 15004
+              mysql_port: 13309
+              grpc_port: 15309
+            - id: 5
+              port: 15005
+              mysql_port: 13310
+              grpc_port: 15310
+            - id: 6
+              port: 15006
+              mysql_port: 13311
+              grpc_port: 15311
     vttablet:
       vars:
         vitess_memory_ratio: 0.6
@@ -74,6 +80,19 @@ all:
           tablets:
             - id: 1001
               keyspace: main
-              shard: 0
+              shard: -80
               pool_size: 500
               transaction_cap: 2000
+              port: 16001
+              grpc_port: 17001
+              mysql_port: 18001
+              mysqld_exporter_port: 9104
+            - id: 2001
+              keyspace: main
+              shard: 80-
+              pool_size: 500
+              transaction_cap: 2000
+              port: 16002
+              grpc_port: 17002
+              mysql_port: 18002
+              mysqld_exporter_port: 9105
