+++ bench_cli/run_benchmark.py
@@ -18,281 +18,24 @@
 
 import os
 import uuid
-import shutil
-import json
-import yaml
-import pysftp
-import abc
 
 import bench_cli.reporting as reporting
 import bench_cli.configuration as configuration
-import bench_cli.packet_vps as packet_vps
-import bench_cli.get_head_hash as get_head_hash
+import bench_cli.task as task
+import bench_cli.task_factory as taskfac
 
-class Task:
-    def __init__(self, report_dir: str, ansible_dir: str, inventory_file: str, source: str, pprof,
-                 create_build_dir: bool = True):
-        self.task_id = uuid.uuid4()
-        self.device_id = 0
-        self.device_ip = ""
-        self.source = source
-        self.commit_hash = ""
-        self.report = None
-        self.pprof = pprof
-        self.report_dir = report_dir
-        self.ansible_dir = ansible_dir
-        self.ansible_build_dir = os.path.join(self.ansible_dir, "build")
-        self.ansible_inventory_file = inventory_file
-        self.ansible_built_inventory_file = self.__build_built_inventory_filename()
-        self.ansible_built_inventory_filepath = os.path.join(self.ansible_build_dir, self.ansible_built_inventory_file)
-
-        if create_build_dir and (
-                not os.path.exists(self.ansible_build_dir) or not os.path.isdir(self.ansible_build_dir)):
-            os.mkdir(self.ansible_build_dir)
-
-    def __build_built_inventory_filename(self):
-        splits = os.path.basename(self.ansible_inventory_file).split('.')
-        return splits[0] + "-" + str(self.task_id) + '.yml'
-
-    def append_state_to_file(self, filepath: str):
-        """
-        Append the task's state to the given file.
-        @param filepath: Filepath used to append the state
-        """
-        curr_state = self.get_state()
-        dump = {'run': []}
-        if os.path.exists(filepath):
-            with open(filepath, 'r') as f:
-                dump = {**dump, **json.load(f)}
-        dump['run'].append(curr_state)
-        with open(filepath, 'w') as outfile:
-            json.dump(dump, outfile)
-
-    def get_state(self):
-        """
-        Returns the task's state.
-
-        @todo: optimize
-        """
-        return {
-            'task_name': self.name(),
-            'run_id': self.task_id.__str__(),
-            'source': self.source,
-            'commit': self.commit_hash,
-            'vps_id': self.device_id,
-            'ip_address': self.device_ip,
-            'inventory_file': os.path.basename(self.ansible_inventory_file)
-        }
-
-    def create_device(self, packet_token, packet_project_id):
-        """
-        Creates the task's Packet Device.
-        @param: packet_token: The packet API token
-        @param: packet_project_id: The packet project ID
-        """
-        vps = packet_vps.create_vps(packet_token, packet_project_id, self.task_id)
-        self.device_id = vps[0]
-        self.device_ip = vps[1]
-        self.append_state_to_file("config-lock.json")
-
-    def delete_device(self, packet_token):
-        """
-        Delete the task's Packet Device.
-        @param: packet_token: The packet API token
-        """
-        packet_vps.delete_vps(packet_token, self.device_id)
-
-    def build_ansible_inventory(self, commit_hash: str):
-        """
-        Create a new Ansible inventory file and build it by appending the task's configuration.
-        The file will be used to run the task on Ansible.
-
-        @param: commit_hash: Commit hash that will be used to run the task
-        """
-        shutil.copy2(self.ansible_inventory_file, self.ansible_built_inventory_filepath)
-        with open(self.ansible_inventory_file, 'r') as invf:
-            invdata = yaml.load(invf, Loader=yaml.FullLoader)
-        self.__recursive_dict(invdata)
-
-        # TODO: handle any commit
-        if commit_hash == 'HEAD':
-            self.commit_hash = get_head_hash.head_commit_hash()
-        invdata["all"]["vars"]["vitess_git_version"] = self.commit_hash
-
-        invdata["all"]["vars"]["pprof_targets"] = None
-        invdata["all"]["vars"]["pprof_args"] = None
-        if self.pprof:
-            invdata["all"]["vars"]["pprof_targets"] = self.pprof["targets"]
-            invdata["all"]["vars"]["pprof_args"] = self.pprof["args"]
-
-        with open(self.ansible_built_inventory_filepath, 'w') as builtf:
-            yaml.dump(invdata, builtf)
-
-    def __recursive_dict(self, inventory_data):
-        for k, _ in inventory_data.items():
-            if isinstance(inventory_data[k], dict) and k == "hosts":
-                inventory_data[k] = self.__recursive_dict_ip(inventory_data[k])
-            elif isinstance(inventory_data[k], dict):
-                inventory_data[k] = self.__recursive_dict(inventory_data[k])
-        return inventory_data
-
-    def __recursive_dict_ip(self, data):
-        for k, _ in data.items():
-            old_key = k
-            data[self.device_ip] = data.pop(old_key)
-        return data
-
-    def __get_remote_task_report(self, echo=False):
-        username = "root"
-        cnopts = pysftp.CnOpts()
-        cnopts.hostkeys = None
-        remote_report_file = self.report_path() + '.bench_report'
-        with pysftp.Connection(host=self.device_ip, username=username, cnopts=cnopts) as sftp:
-            if echo:
-                print("Connection succesfully stablished ... ")
-            remote_file_path = '/tmp/' + self.name() + '.json'
-            sftp.get(remote_file_path, remote_report_file)
-        with open(remote_report_file, 'r') as f:
-            task_report = json.load(f)
-        return task_report
-
-    def save_report(self):
-        """
-        Save the task's state to a report file.
-        """
-        task_report = self.__get_remote_task_report()
-        if len(task_report) == 0:
-            return
-        self.report = {**self.get_state(), 'results': task_report[0]}
-        with open(self.report_path(), 'w') as f:
-            json.dump(self.report, f)
-
-    def clean_up(self):
-        """
-        Removes the ansible_built_inventory_file of the file system.
-        """
-        os.remove(self.ansible_built_inventory_file)
-
-    @abc.abstractmethod
-    def run(self, script_path: str):
-        pass
-
-    @abc.abstractmethod
-    def name(self) -> str:
-        pass
-
-    @abc.abstractmethod
-    def report_path(self, base: str = None) -> str:
-        pass
-
-    @abc.abstractmethod
-    def table_name(self) -> str:
-        pass
-
-
-class OLTP(Task):
-    def name(self) -> str:
-        """
-        Returns the task's name
-        """
-        return 'oltp'
-
-    def run(self, script_path: str):
-        """
-        Runs the task.
-
-        @param: script_path: Path to the Ansible script's directory.
-
-        @todo: Use the Ansible Galaxy API
-        """
-        os.system(os.path.join(script_path, "run-oltp") + ' ' + self.ansible_built_inventory_filepath + ' ' + self.ansible_dir)
-
-    def report_path(self, base: str = None) -> str:
-        """
-        Returns the path of the task report directory.
-
-        @param: base: Folder to use as base for the report directory
-        """
-        if base is not None:
-            return os.path.join(base, "oltp_v2.json")
-        return os.path.join(self.report_dir, "oltp_v2.json")
-
-    def table_name(self) -> str:
-        """
-        Returns the task's table name
-        """
-        return "OLTP"
-
-
-class TPCC(Task):
-    def name(self) -> str:
-        """
-        Returns the task's name
-        """
-        return 'tpcc'
-
-    def run(self, script_path: str):
-        """
-        Runs the task.
-
-        @param: script_path: Path to the Ansible script's directory.
-
-        @todo: Use the Ansible Galaxy API
-        """
-        os.system(os.path.join(script_path, "run-tpcc") + ' ' + self.ansible_built_inventory_filepath + ' ' + self.ansible_dir)
-
-    def report_path(self, base: str = None) -> str:
-        """
-        Returns the path of the task report directory.
-
-        @param: base: Folder to use as base for the report directory
-        """
-        if base is not None:
-            return os.path.join(base, "tpcc_v2.json")
-        return os.path.join(self.report_dir, "tpcc_v2.json")
-
-    def table_name(self) -> str:
-        """
-        Returns the task's table name
-        """
-        return "TPCC"
-
-
-class TaskFactory:
-    def __init__(self):
-        pass
-
-    def create_task(self, task_name, report_dir, ansible_dir, inventory_file, source, pprof) -> Task:
-        """
-        Create a task children based on the given task_name.
-        The task created can either be "oltp" (OLTP) or "tpcc" (TPCC).
-
-        @param: task_name
-        @param: report_dir: Path to task's report directory
-        @param: ansible_dir: Path to the Ansible directory to use
-        @param: inventory_file: Filename of the inventory to use
-        @param: source: The task's source
-        @param: pprof: The pprof configuration of the task
-        """
-        if task_name == "oltp":
-            return OLTP(report_dir, ansible_dir, inventory_file, source, pprof)
-        elif task_name == "tpcc":
-            return TPCC(report_dir, ansible_dir, inventory_file, source, pprof)
-
-
-# ------------------------------------------------------ Runs benchmark tasks ---------------------------------------------------------
 
 class BenchmarkRunner:
     def __init__(self, config: configuration.Config, echo=False):
         self.runner_id = uuid.uuid4()
         self.config = config
-        self.tasks = self.__instantiate_tasks__()
+        self.tasks = self.__instantiate_tasks()
         if echo:
             print('Runner ' + self.runner_id.__str__() + ' created.')
 
-    def __instantiate_tasks__(self) -> [Task]:
+    def __instantiate_tasks(self) -> [task.Task]:
         tasks = []
-        task_factory = TaskFactory()
+        task_factory = taskfac.TaskFactory()
         for task_name in self.config.tasks:
             tasks.append(task_factory.create_task(task_name, self.config.tasks_reports_dir,
                                                   self.config.ansible_dir,
@@ -311,5 +54,6 @@ class BenchmarkRunner:
             task.build_ansible_inventory(self.config.commit)
             task.run(self.config.tasks_scripts_dir)
             task.save_report()
+            task.download_remote_pprof_folder()
             reporting.save_to_mysql(self.config, task.report, task.table_name())
             reporting.send_slack_message(self.config.slack_api_token, self.config.slack_channel, task.report_path())
